+++ colossalai/shardformer/layer/qkv_fused_linear.py
@@ -265,6 +265,14 @@ class GPT2FusedLinearConv1D_Col(ParallelModule):
                 f'Expected only one process group, got {len(process_group)}.'
             process_group = process_group[0]
 
+        tp_size = dist.get_world_size(process_group)
+        if out_features < tp_size:
+            return module
+
+        if out_features % tp_size != 0:
+            raise ValueError(
+                f"The size of out_features:{out_features} is not integer multiples of tensor parallel size: {tp_size}!")
+
         linear_1d = GPT2FusedLinearConv1D_Col(in_features=in_features,
                                               out_features=out_features,
                                               bias=bias,
@@ -419,6 +427,14 @@ class GPT2FusedLinearConv1D_Row(ParallelModule):
             assert len(process_group) == 1, \
                 f'Expected only one process group, got {len(process_group)}.'
             process_group = process_group[0]
+
+        tp_size = dist.get_world_size(process_group)
+        if in_features < tp_size:
+            return module
+
+        if in_features % tp_size != 0:
+            raise ValueError(
+                f"The size of in_features:{in_features} is not integer multiples of tensor parallel size: {tp_size}!")
 
         linear_1d = GPT2FusedLinearConv1D_Row(in_features=in_features,
                                               out_features=out_features,
